Revert "Merge remote-tracking branch 'freebsd/stable/11' into freebsd11"
This reverts commit ecba3d8cfbbacf161276d834187410d90ae4d627, reversing
changes made to 41e6a40a2019cc527dc3506fa731c83175340d8a.

+++ sys/kern/sched_ule.c
@@ -1662,11 +1662,7 @@ sched_pctcpu_update(struct td_sched *ts, int run)
 {
 	int t = ticks;
 
-	/*
-	 * The signed difference may be negative if the thread hasn't run for
-	 * over half of the ticks rollover period.
-	 */
-	if ((u_int)(t - ts->ts_ltick) >= SCHED_TICK_TARG) {
+	if (t - ts->ts_ltick >= SCHED_TICK_TARG) {
 		ts->ts_ticks = 0;
 		ts->ts_ftick = t - SCHED_TICK_TARG;
 	} else if (t - ts->ts_ftick >= SCHED_TICK_MAX) {
